Fix a case that potentially returns not enough messages
If the frontend requests a Newest - 5 for a single partition there was a chance that the FetchMessages() method would return earlier than desired because with multiple workers the right order within each partition is no longer given. It is possible that the receiving routine would receive message with the last message with offset 100 before it has received the message with offset 98. It would then return early because it knows that 100 is the high watermark. To avoid this case we'll use worker count = 1 where reasonable to avoid it as much as possible.