helm: git credentials secret
This commit adds support for a embedding a github basic auth password
as a kubernetes `secret` in the helm chart.

Before this change, when setting up a git repository as a protobuf
registry forprotobuf deserialization, you could provide basic auth
credentials through the helm chart. However you could not provide the
secret to the helm chart without exposing it in a `configmap` resource.

After this change, you can now provide it to the helm chart via:

```yaml
secret:
    kafka:
        protobufGitBasicAuthPassword: <MYPASSWORD>
```

The change follows the pattern already present in the helm chart for
console secrets:
1. add the value to the `secret` resource
2. map it the value from the `secret` to an environment variable using
   the `deployment` resource.
3. the environment variable name used is the config path that you would
   typically use in the yaml with the `.`s replaced by `_`s: i.e.
   `kafka.protobuf.git.basicAuth.Password`,
   becomes `KAFKA_PROTOBUF_GIT_BASICAUTH_PASSWORD`.

By doing this, its now possible for the user of the helm chart to
embed the github basic auth password in the helm chart
as a kubernetes `secret`.