proto: allow multiple import paths
add new config parameter proto.importPaths.
if import paths are configured, these paths are all treated as
roots. if not configured, the already existing behavior is preserved.

in addition, if git.basePath is empty / "." (default), the dot from file
extensions is dropped, preventing correctly parsing protos. the dot is
now not stripped anymore.