backend: refresh aws creds when using iam auth after token expiration
Currently, when AWS IAM auth fetches token, it only does so once on startup and the same token is reused. This change introduces refresh mechanism by retrieving the token asynchronously when it's needed by the kafka auth system.

Token is internally cached and reused until it expires. Once it expires, a new token is fetched and used.