rpcn: use a generic resources field instead of requests (k8s semantics)
in serverless, users will always see a certain amount of resources in the UI e.g. 100 millicores, but under the hood we may decide to set a "k8s request" e.g. of 10 millicores if CPU is mostly unused on average on some nodes.

So the value the user configures in serverless is basically the "limit", not the "request" (which is on us to tune). Hence, better calling it just "resources".