servicer: add Create/DeleteSRACLs and a probe
This probe (CheckSchemaRegistryACLSupport)
is our way to identify if Schema Registry
ACLs are usable in the cluster. The API
doesn't expose a method for that, so we
rely on sending a faulty ACL Create request,
the API will validate first and we respond
accordingly.

If it's a 404: then it's an older cluster
or a cluster that doesn't support SR ACLs

If it's a 403: then it might be that either
the user is not allowed to use the endpoint
or you have an invalid license, so we need
to filter by the word 'license'.