refactor MCPServer resource
do not talk about config files, but about MCP tools.
this makes the resource more natural to use:

Declare tools and their YAMLs.
Also add field "compontent_type"; we need to know what type it
is to mount it into the correct folder.

this takes away the burden to understand the folder structure.
in addition, it makes the folder structure an implementation detail, not
part of the API.