Categorize and sort topic configuration options (#1929)
The topic config options we know about are significantly
out of date. This updates the set of options that we know
about.

Currently the groups on the frontend are sorted arbitrarily
by the order that we receive the current options in. This
update allows us to specify a group ordering to help push
common configuration groups towards the top of the page.