Increase timeout and use filtering in tests
The e2e enterprise tests were timing out after the default (30s)
timeout, and, on several retries, some of the values we were looking for
were getting pushed to the second page of results.

This bumps the timeout to 60s and enters style principal name in the
filter box before checking that t in the filter box before checking that
the entry exists.