Loosen validation on secret update values (#1962)
Loosen validation on secret update values

In order to support updating labels/scopes on secrets without
needing to set a new value, we need to remove some validation from the
proto definition. This is the middle commit between a console-enterprise
commit to support metadata-only updates and an update to the zod
validation on console's frontend.