frontend: enable scanner rules
https://biomejs.dev/internals/architecture/#scanner

Scanner
Biome has a scanner that is responsible for crawling the file system to extract important metadata about projects. Specifically, there are three ways in which the scanner is used:

To discover nested biome.json/biome.jsonc files in monorepos.
To discover nested .gitignore files if the vcs.useIgnoreFile setting is enabled.
To index package.json manifests as well as source files in a project if any rules from the project domain are enabled.